Folder level changes only.
No changes to actual code except last item.

Deleted these notebooks/files:
FindingACorrelation.ipynb
FirstGeoPandasplot.ipynb
ETLForElectionAndVaccinationData-Original.py
CovidRollingAveragesAndMaskData-With Delta variant - Copy.ipynb

Corrected the name on this one so it may appear as if the old one was deleted
and this is a new file:
CodeForPresidentialElectionsAndCovid-Interactive.ipynb

The only actual code changes were to this notebook.
Brought back the code that was overwritten:
UrbanRuralConnectionV1.ipynb